src/fitnessd_client/requirements.txt
requests
gooey
python-dateutil